requirements.txt
fastapi
uvicorn[standard]
requests
aiohttp
sqlalchemy
pillow
pydantic-settings
pytest-asyncio
pytest
httpx